# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "싱송생송 API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/v1/recommend/recommendation/functionCalling": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "LLM의 사용자 입력을 토대로 추천된 노래를 반환합니다. 10개의 노래를 반환합니다",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recommendation"
                ],
                "summary": "LLM으로 검색하기",
                "parameters": [
                    {
                        "description": "인풋",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.LlmRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "성공",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/pkg.BaseResponseStruct"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handler.FunctionCallingResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/v2/recommend/recommendation/functionCallingWithTypes": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "LLM의 사용자 입력을 토대로 추천된 노래를 반환합니다. 20개의 노래를 반환합니다",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recommendation"
                ],
                "summary": "LLM으로 검색하기",
                "parameters": [
                    {
                        "description": "인풋",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.LlmRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "성공",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/pkg.BaseResponseStruct"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handler.FunctionCallingWithTypesResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.FunctionCallingDetailResponse": {
            "type": "object",
            "properties": {
                "album": {
                    "type": "string"
                },
                "commentCount": {
                    "type": "integer"
                },
                "isKeep": {
                    "type": "boolean"
                },
                "isLive": {
                    "type": "boolean"
                },
                "isMr": {
                    "type": "boolean"
                },
                "keepCount": {
                    "type": "integer"
                },
                "lyricsYoutubeLink": {
                    "type": "string"
                },
                "melonLink": {
                    "type": "string"
                },
                "singerName": {
                    "type": "string"
                },
                "songId": {
                    "type": "integer"
                },
                "songName": {
                    "type": "string"
                },
                "songNumber": {
                    "type": "integer"
                },
                "tjYoutubeLink": {
                    "type": "string"
                }
            }
        },
        "handler.FunctionCallingResponse": {
            "type": "object",
            "properties": {
                "songs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.FunctionCallingDetailResponse"
                    }
                }
            }
        },
        "handler.FunctionCallingWithTypesResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "songs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.FunctionCallingDetailResponse"
                    }
                }
            }
        },
...
        "handler.LlmRequest": {
            "type": "object",
            "properties": {
                "userInput": {
                    "type": "string"
                }
            }
        },
...
        "pkg.BaseResponseStruct": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}